backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain-community>=0.3.29",
    "langchain-text-splitters>=0.3.11",
    "langchain[groq]>=0.3.27",
    "langgraph>=0.6.7",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
]
